Fix bug when creating translation .properties files.
The dictionary with the translations was not cleared after writing the file
(and changing the language) the translations of the previously processed
language were "inherited" in the following one.

This could cause incorrect strings to be used in a language that did not have
translations (such as en_US).
(grafted from 32f4f1001561123a1cf86b5b6365a03a25f67f30)